+++ IronSource/src/main/java/com/mopub/mobileads/IronSourceAdapterConfiguration.java
@@ -27,7 +27,7 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM_WITH_THROWABLE;
 
 public class IronSourceAdapterConfiguration extends BaseAdapterConfiguration {
-    public static final String IRONSOURCE_ADAPTER_VERSION = "500";
+    public static final String IRONSOURCE_ADAPTER_VERSION = "510";
     public static final String DEFAULT_INSTANCE_ID = "0";
 
     private static final String ADAPTER_NAME = IronSourceAdapterConfiguration.class.getSimpleName();
@@ -39,7 +39,7 @@ public class IronSourceAdapterConfiguration extends BaseAdapterConfiguration {
     private static final String INTERSTITIAL_KEY = "interstitial";
     private static final String MEDIATION_TYPE = "mopub";
     private static final String REWARDEDVIDEO_KEY = "rewardedvideo";
-    
+
     @NonNull
     @Override
     public String getAdapterVersion() {
@@ -85,36 +85,30 @@ public class IronSourceAdapterConfiguration extends BaseAdapterConfiguration {
 
         boolean networkInitializationSucceeded = false;
 
+
         synchronized (IronSourceAdapterConfiguration.class) {
             try {
                 if (configuration != null) {
-                    if (!(context instanceof Activity)) {
-                        MoPubLog.log(CUSTOM, ADAPTER_NAME, "IronSource's initialization via " +
-                                ADAPTER_NAME + " not started. An Activity Context is needed.");
-                        return;
-                    }
 
                     final String appKey = configuration.get(APPLICATION_KEY);
 
-                    if (TextUtils.isEmpty(appKey)) {
+                    if (!TextUtils.isEmpty(appKey)) {
+                        IronSource.AD_UNIT[] adUnitsToInitList = getIronSourceAdUnitsToInitList(context, configuration);
+                        initIronSourceSDK(context, appKey, adUnitsToInitList);
+                        networkInitializationSucceeded = true;
+                    } else {
                         MoPubLog.log(CUSTOM, "IronSource's initialization not" +
                                 " started. Ensure ironSource's " + APPLICATION_KEY +
                                 " is populated on the MoPub dashboard.");
-                    } else {
-                        IronSource.AD_UNIT[] adUnitsToInitList = getIronSourceAdUnitsToInitList(context, configuration);
-                        initIronSourceSDK((Activity) context, appKey, adUnitsToInitList);
-                        networkInitializationSucceeded = true;
                     }
-
                 } else {
                     MoPubLog.log(CUSTOM, ADAPTER_NAME, "IronSource's initialization via " +
                             ADAPTER_NAME + " not started. No configuration information present to initialize." +
-                            "Make sure you pass applicationKey on MoPub UI " +
-                            "or via Mediated Network Configuration during MoPub initialization.");
+                            "Make sure to pass in ironSource appKey parameter to MoPub initialization via network configuration");
                 }
             } catch (Exception e) {
                 MoPubLog.log(CUSTOM_WITH_THROWABLE, "Initializing ironSource has encountered " +
-                        "an exception.", e);
+                        "an exception. ", e);
             }
         }
 
@@ -123,7 +117,7 @@ public class IronSourceAdapterConfiguration extends BaseAdapterConfiguration {
                     MoPubErrorCode.ADAPTER_INITIALIZATION_SUCCESS);
         } else {
             MoPubLog.log(CUSTOM, ADAPTER_NAME, "IronSource's initialization via " +
-                    ADAPTER_NAME + " failed. Will attempt to initialize on the first ad request to ironSource.");
+                    ADAPTER_NAME + " failed.");
             listener.onNetworkInitializationFinished(IronSourceAdapterConfiguration.class,
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
         }
@@ -138,10 +132,11 @@ public class IronSourceAdapterConfiguration extends BaseAdapterConfiguration {
         }
     }
 
-    public static void initIronSourceSDK(Activity activity, String appKey, IronSource.AD_UNIT[] adUnitsToInitList) {
-        IronSource.setMediationType(MEDIATION_TYPE + IRONSOURCE_ADAPTER_VERSION
+    public static void initIronSourceSDK(@NonNull Context context, @NonNull String appKey, IronSource.AD_UNIT[] adUnitsToInitList) {
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "IronSource initialization started with appKey: " + appKey);
+                IronSource.setMediationType(MEDIATION_TYPE + IRONSOURCE_ADAPTER_VERSION
                 + "SDK" + getMoPubSdkVersion());
-        IronSource.initISDemandOnly(activity, appKey, adUnitsToInitList);
+        IronSource.initISDemandOnly(context, appKey, adUnitsToInitList);
     }
 
     /**