+++ IronSource/src/main/java/com/mopub/mobileads/IronSourceRewardedVideo.java
@@ -42,11 +42,9 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
     // Configuration keys
     private static final String APPLICATION_KEY = "applicationKey";
     private static final String INSTANCE_ID_KEY = "instanceId";
-    private static final String MEDIATION_TYPE = "mopub";
     private static final String ADAPTER_NAME = IronSourceRewardedVideo.class.getSimpleName();
 
     // Network identifier of ironSource
-    @NonNull
     private String mInstanceId = IronSourceAdapterConfiguration.DEFAULT_INSTANCE_ID;
 
     @NonNull
@@ -59,7 +57,7 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
     private IronSourceAdapterConfiguration mIronSourceAdapterConfiguration;
 
     /**
-     * Mopub API
+     * MoPub API
      */
 
     public IronSourceRewardedVideo() {
@@ -117,8 +115,19 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
                 mInstanceId = instanceId;
             }
 
-            initIronSourceSDK(launcherActivity, applicationKey, extras);
-            return true;
+            final Context context = launcherActivity.getApplicationContext();
+
+            if (context != null) {
+                initIronSourceSDK(context, applicationKey, extras);
+                return true;
+            } else {
+                MoPubLog.log(CUSTOM, ADAPTER_NAME, "ironSource Interstitial failed to initialize." +
+                        "Application Context obtained by Activity launching this interstitial is null.");
+                if (mLoadListener != null) {
+                    mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+                }
+                return false;
+            }
 
         } catch (Exception e) {
             MoPubLog.log(CUSTOM_WITH_THROWABLE, e);
@@ -131,12 +140,12 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
         }
     }
 
-    private void initIronSourceSDK(Activity activity, String applicationKey, Map<String, String> extras) {
+    private void initIronSourceSDK(Context context, String applicationKey, Map<String, String> extras) {
         MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "ironSource Rewarded Video initialization is called with applicationKey: " + applicationKey);
         IronSource.setISDemandOnlyRewardedVideoListener(this);
 
-        IronSource.AD_UNIT[] adUnitsToInit = mIronSourceAdapterConfiguration.getIronSourceAdUnitsToInitList(activity, extras);
-        IronSourceAdapterConfiguration.initIronSourceSDK(activity, applicationKey, adUnitsToInit);
+        IronSource.AD_UNIT[] adUnitsToInit = mIronSourceAdapterConfiguration.getIronSourceAdUnitsToInitList(context, extras);
+        IronSourceAdapterConfiguration.initIronSourceSDK(context, applicationKey, adUnitsToInit);
     }
 
     @Override
@@ -145,6 +154,15 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
         Preconditions.checkNotNull(adData);
 
         setAutomaticImpressionAndClickTracking(false);
+
+        if (!(context instanceof Activity)) {
+            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(), MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Failed to load rewarded video as ironSource requires an Activity context.");
+            if (mLoadListener != null) {
+                mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            }
+            return;
+        }
 
         final Map<String, String> extras = adData.getExtras();
 
@@ -166,9 +184,9 @@ public class IronSourceRewardedVideo extends BaseAd implements ISDemandOnlyRewar
 
         if(!TextUtils.isEmpty(adMarkup)) {
             MoPubLog.log(CUSTOM, ADAPTER_NAME, "ADM field is populated. Will make Advanced Bidding request.");
-            IronSource.loadISDemandOnlyRewardedVideoWithAdm(mInstanceId, adMarkup);
+            IronSource.loadISDemandOnlyRewardedVideoWithAdm((Activity) context, mInstanceId, adMarkup);
         } else {
-            IronSource.loadISDemandOnlyRewardedVideo(mInstanceId);
+            IronSource.loadISDemandOnlyRewardedVideo((Activity) context, mInstanceId);
         }
     }