+++ IronSource/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+ * 7.1.3.0
+    * This version of the adapters has been certified with ironSource 7.1.3 and MoPub SDK 5.16.2.
+    * Add support for ironSource SDK to initialize with `Context` instead of `Activity`.
+
  * 7.1.0.2.0
     * This version of the adapters has been certified with ironSource 7.1.0.2 and MoPub SDK 5.15.0.